database: support disabling cgo by removing sqlite support
