+++ .github/workflows/cgo.yml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   build:
-    name: Go Build
+    name: Go Build (CGO)
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
@@ -17,7 +17,7 @@ jobs:
     - name: Set up Go 1.x
       uses: actions/setup-go@v2
       with:
-        go-version: ^1.14
+        go-version: ^1.16
       id: go
 
     - name: Check out code into the Go module directory
