+++ database/migrator.go
@@ -1,3 +1,6 @@
+// Copyright 2020 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
 package database
 
 import (
